requirements.txt
pycryptodome
pyotp
requests
